debugfs: Suppress build warnings
This patch silences the following warning:
dump.c:214: warning: format ‘%lu’ expects type ‘long unsigned int’, but argument 3 has type ‘__u64’

+++ debugfs.ocfs2/commands.c
@@ -494,7 +494,7 @@ static errcode_t find_block_offset(ocfs2_filesys *fs,
 		ret = ocfs2_read_extent_block(fs, rec->e_blkno, buf);
 		if (ret) {
 			com_err(gbls.cmd, ret, "while reading extent %"PRIu64,
-				rec->e_blkno);
+				(uint64_t)rec->e_blkno);
 			goto bail;
 		}
